+++ example/src/App.svelte
@@ -2,8 +2,4 @@
 	import Typewriter from 'Typewriter.svelte'
 </script>
 
-<Typewriter loopRandom>
-	<p>Lorem</p>
-  <p>Ipsum</p>
-  <p>Dolor</p>
-</Typewriter>
+<Typewriter loopRandom>Hello World!</Typewriter>